refactor: 플레이어의 상태 확인 책임을 State에서 Player로 이동
