+++ src/main/java/blackjack/domain/player/Player.java
@@ -1,5 +1,6 @@
 package blackjack.domain.player;
 
+import blackjack.domain.Rule;
 import blackjack.domain.card.Card;
 import blackjack.domain.card.Cards;
 import java.util.ArrayList;
@@ -8,6 +9,7 @@ import java.util.regex.Pattern;
 public abstract class Player {
 
     private static final Pattern NON_SPECIAL_CHARACTERS = Pattern.compile("^[0-9a-zA-Zㄱ-ㅎㅏ-ㅣ가-힣\\s]*$");
+    private static final int BLACKJACK_CARDS_SIZE = 2;
 
     private final String name;
     private final Cards cards;
@@ -43,7 +45,7 @@ public abstract class Player {
     }
 
     public boolean isBust() {
-        return State.from(cards) == State.BUST;
+        return cards.calculateTotalScore() > Rule.WINNING_SCORE.getValue();
     }
 
     public boolean isNotBust() {
@@ -51,13 +53,13 @@ public abstract class Player {
     }
 
     public boolean isBlackjack() {
-        return State.from(cards) == State.BLACKJACK;
+        return cards.getSize() == BLACKJACK_CARDS_SIZE &&
+                cards.calculateTotalScore() == Rule.WINNING_SCORE.getValue();
     }
 
     public boolean isNotBlackjack() {
         return !isBlackjack();
     }
-
 
     public String getName() {
         return name;
